terminando swagger

+++ 85ESE/product_api/data/products.go
@@ -6,35 +6,22 @@ import (
 
 var ErrProductNotFound = fmt.Errorf("Product not found")
 
+// Produto representa um item disponível no sistema.
+// swagger:model
 type Product struct {
-	// the id for the product
-	//
-	// required: false
-	// min: 1
-	ID int `json:"id"` // Unique identifier for the product
+	// ID do produto
+	ID int `json:"id"`
 
-	// the name for this poduct
-	//
-	// required: true
-	// max length: 255
+	// Nome do produto
 	Name string `json:"name" validate:"required"`
 
-	// the description for this poduct
-	//
-	// required: false
-	// max length: 10000
+	// Descrição do produto
 	Description string `json:"description"`
 
-	// the price for the product
-	//
-	// required: true
-	// min: 0.01
-	Price float32 `json:"price" validate:"required,gt=0"`
+	// Preço do produto
+	Price float64 `json:"price" validate:"required,gt=0"`
 
-	// the SKU for the product
-	//
-	// required: true
-	// pattern: [a-z]+-[a-z]+-[a-z]+
+	// SKU do produto
 	SKU string `json:"sku" validate:"sku"`
 }
